pyproject.toml
[project]
name = "ai-assistant-for-work-chats"
version = "0.1.0"
description = "first to encounter"
authors = [
    {name = "trueforme"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic (>=1.16.5,<2.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "fastapi (>=0.118.0,<0.119.0)",
    "uvicorn[standard] (>=0.37.0,<0.38.0)",
    "psycopg[binary] (>=3.2.10,<4.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[dependency-groups]
dev = [
    "pytest (>=8.4.2,<9.0.0)",
    "ruff (>=0.13.3,<0.14.0)",
    "black (>=25.9.0,<26.0.0)",
    "mypy (>=1.18.2,<2.0.0)"
]
